requirements.txt
albumentations==1.3.1
einops==0.6.1
gym==0.26.2
imageio==2.28.1
ipython==8.18.1
lime==0.2.0.1
matplotlib==3.7.1
numpy==1.23.5
opencv-python==4.8.0.74
pandas==2.0.1
Pillow>=10.3.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy
torch
torchtext
torchvision
tqdm
ipywidgets
transformers
flask